♻️ Add a hook for reused progress logic
And so much less code :D
