+++ apps/storybook-react/stories/Progress Indicators/Circular.stories.tsx
@@ -1,9 +1,10 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import {
   CircularProgress,
   CircularProgressProps,
 } from '@equinor/eds-core-react'
 import { Meta, Story } from '@storybook/react'
+import { useProgress } from './hooks/useProgress'
 
 export default {
   title: 'Components/Progress Indicators/Circular',
@@ -12,26 +13,8 @@ export default {
 
 export const Default: Story<CircularProgressProps> = (args) => {
   const { value = 0, variant } = args
-  let controllableValue: number | null
-  variant === 'indeterminate'
-    ? (controllableValue = null)
-    : (controllableValue = value)
-  const [progress, setProgress] = useState(controllableValue)
-  useEffect(() => {
-    const timer = setInterval(() => {
-      setProgress((oldProgress) => {
-        if (oldProgress === 100) {
-          return 100
-        }
-        const diff = Math.random() * 10
-        return Math.min(oldProgress + diff, 100)
-      })
-    }, 500)
+  const progress = variant === 'indeterminate' ? null : useProgress(value)
 
-    return () => {
-      clearInterval(timer)
-    }
-  }, [])
   return <CircularProgress {...args} value={progress} />
 }
 
@@ -40,23 +23,7 @@ export const Indeterminate: Story<CircularProgressProps> = () => {
 }
 
 export const Determinate: Story<CircularProgressProps> = () => {
-  const [progress, setProgress] = useState(0)
-
-  useEffect(() => {
-    const timer = setInterval(() => {
-      setProgress((oldProgress) => {
-        if (oldProgress === 100) {
-          return 100
-        }
-        const diff = Math.random() * 10
-        return Math.min(oldProgress + diff, 100)
-      })
-    }, 500)
-
-    return () => {
-      clearInterval(timer)
-    }
-  }, [])
+  const progress = useProgress(0)
   return <CircularProgress variant="determinate" value={progress} />
 }
