+++ apps/storybook-react/stories/Progress Indicators/Star.stories.tsx
@@ -1,6 +1,7 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import { StarProgress, StarProgressProps } from '@equinor/eds-core-react'
 import { Meta, Story } from '@storybook/react'
+import { useProgress } from './hooks/useProgress'
 
 export default {
   title: 'Components/Progress Indicators/Star',
@@ -9,49 +10,15 @@ export default {
 
 export const Default: Story<StarProgressProps> = (args) => {
   const { value = 0, variant } = args
-  let controllableValue: number | null
-  variant === 'indeterminate'
-    ? (controllableValue = null)
-    : (controllableValue = value)
-  const [progress, setProgress] = useState(controllableValue)
-  useEffect(() => {
-    const timer = setInterval(() => {
-      setProgress((oldProgress) => {
-        if (oldProgress === 100) {
-          return 100
-        }
-        const diff = Math.random() * 10
-        return Math.min(oldProgress + diff, 100)
-      })
-    }, 500)
+  const progress = variant === 'indeterminate' ? null : useProgress(value)
 
-    return () => {
-      clearInterval(timer)
-    }
-  }, [])
   return <StarProgress value={progress} {...args} />
 }
 
 export const Indeterminate: Story<StarProgressProps> = () => <StarProgress />
 
 export const Determinate: Story<StarProgressProps> = () => {
-  const [progress, setProgress] = useState(0)
-
-  useEffect(() => {
-    const timer = setInterval(() => {
-      setProgress((oldProgress) => {
-        if (oldProgress === 100) {
-          return 100
-        }
-        const diff = Math.random() * 10
-        return Math.min(oldProgress + diff, 100)
-      })
-    }, 500)
-
-    return () => {
-      clearInterval(timer)
-    }
-  }, [])
+  const progress = useProgress(0)
   return <StarProgress value={progress} variant="determinate" />
 }
